+++ outputs-kv.tf
@@ -20,3 +20,8 @@ output "keyvault_uri" {
   description = "URI of the Key Vault."
   value       = module.keyvault.key_vault_uri
 }
+
+output "key_vault_hsm_uri" {
+  description = "The URI of the Key Vault Managed Hardware Security Module, used for performing operations on keys."
+  value       = module.keyvault.key_vault_hsm_uri
+}